Extract knob story
change to themed
